게시글 작성 부분 분리

+++ src/main/java/com/dita/metapilot/post/dto/HashtagDto.java
@@ -13,6 +13,6 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 public class HashtagDto {
-    private int hashtagId;
+    private long hashtagId;
     private String content;
 }
